+++ shaders/program/composite11.glsl
@@ -2,30 +2,43 @@ in_out vec2 texcoord;
 
 
 
+/*
+Generated using this rust code:
+let count = 8;
+let mut total = 0.0;
+for x in -count..=count {
+	for y in -count..=count {
+		let fx = x as f32 / count as f32 * 1.73081845045;
+		let fy = x as f32 / count as f32 * 1.73081845045;
+		let vx = f32::consts::E.powf(fx * fx * -1.);
+		let vy = f32::consts::E.powf(fy * fy * -1.);
+		if y == 0 { println!("{vx}"); }
+		total += vx * vy;
+	}
+}
+println!("{}", total);
+*/
+
 #if BLOOM_QUALITY == 1
-	#define SAMPLE_COUNT 3
-	const float stepAmount = 0.4 * BLOOM_SIZE * 0.1;
-	const float sampleWeights[SAMPLE_COUNT] = float[SAMPLE_COUNT] (0.852, 0.527, 0.23);
-	const float weightsTotal = 4.218;
+	#define HALF_SAMPLE_COUNT 3
+	#define SAMPLE_COUNT (HALF_SAMPLE_COUNT * 2 + 1)
+	const float sampleWeights[SAMPLE_COUNT] = float[SAMPLE_COUNT] (0.050000004, 0.2640976, 0.71687114, 1, 0.71687114, 0.2640976, 0.050000004);
+	const float weightsTotal = 15.206126;
 #elif BLOOM_QUALITY == 2
-	#define SAMPLE_COUNT 4
-	const float stepAmount = 0.32 * BLOOM_SIZE * 0.1;
-	const float sampleWeights[SAMPLE_COUNT] = float[SAMPLE_COUNT] (0.902, 0.664, 0.398, 0.194);
-	const float weightsTotal = 5.316;
+	#define HALF_SAMPLE_COUNT 4
+	#define SAMPLE_COUNT (HALF_SAMPLE_COUNT * 2 + 1)
+	const float sampleWeights[SAMPLE_COUNT] = float[SAMPLE_COUNT] (0.050000004, 0.185426, 0.4728708, 0.8292503, 1, 0.8292503, 0.4728708, 0.185426, 0.050000004);
+	const float weightsTotal = 26.066614;
 #elif BLOOM_QUALITY == 3
-	#define SAMPLE_COUNT 6
-	const float sampleWeights[SAMPLE_COUNT] = float[SAMPLE_COUNT] (0.9394131, 0.7788008, 0.56978285, 0.36787945, 0.2096114, 0.10539923);
-	const float weightsTotal = 6.9417734;
+	#define HALF_SAMPLE_COUNT 6
+	#define SAMPLE_COUNT (HALF_SAMPLE_COUNT * 2 + 1)
+	const float sampleWeights[SAMPLE_COUNT] = float[SAMPLE_COUNT] (0.050000004, 0.12488407, 0.2640976, 0.4728708, 0.71687114, 0.9201535, 1, 0.9201535, 0.71687114, 0.4728708, 0.2640976, 0.12488407, 0.050000004);
+	const float weightsTotal = 56.47294;
 #elif BLOOM_QUALITY == 4
-	#define SAMPLE_COUNT 8
-	const float stepAmount = 0.17 * BLOOM_SIZE * 0.1;
-	const float sampleWeights[SAMPLE_COUNT] = float[SAMPLE_COUNT] (0.971, 0.891, 0.771, 0.63, 0.486, 0.353, 0.243, 0.157);
-	const float weightsTotal = 10.004;
-#elif BLOOM_QUALITY == 5
-	#define SAMPLE_COUNT 12
-	const float stepAmount = 0.115 * BLOOM_SIZE * 0.1;
-	const float sampleWeights[SAMPLE_COUNT] = float[SAMPLE_COUNT] (0.986, 0.948, 0.888, 0.81, 0.718, 0.621, 0.523, 0.429, 0.343, 0.266, 0.212, 0.149);
-	const float weightsTotal = 14.786;
+	#define HALF_SAMPLE_COUNT 8
+	#define SAMPLE_COUNT (HALF_SAMPLE_COUNT * 2 + 1)
+	const float sampleWeights[SAMPLE_COUNT] = float[SAMPLE_COUNT] (0.050000004, 0.10090181, 0.185426, 0.31030244, 0.4728708, 0.6562097, 0.8292503, 0.9542703, 1, 0.9542703, 0.8292503, 0.6562097, 0.4728708, 0.31030244, 0.185426, 0.10090181, 0.050000004);
+	const float weightsTotal = 98.45921;
 #endif
 
 
@@ -34,7 +47,15 @@ in_out vec2 texcoord;
 
 const bool colortex4MipmapEnabled = true;
 
-#include "/utils/depth.glsl"
+void doBloomTile(inout vec3 bloomColor, vec2 stepAmount) {
+	for (int x = -HALF_SAMPLE_COUNT; x <= HALF_SAMPLE_COUNT; x++) {
+		for (int y = -HALF_SAMPLE_COUNT; y <= HALF_SAMPLE_COUNT; y++) {
+			vec2 offset = stepAmount * (vec2(x, y) / HALF_SAMPLE_COUNT);
+			float weight = sampleWeights[x + HALF_SAMPLE_COUNT] * sampleWeights[y + HALF_SAMPLE_COUNT];
+			bloomColor += texture2DLod(BLOOM_TEXTURE, texcoord + offset, 2).rgb * weight;
+		}
+	}
+}
 
 void main() {
 	#if HORROR_MODE == 1
@@ -42,20 +63,16 @@ void main() {
 		return;
 	#endif
 	
-	const int bloomIntScale = 1 << BLOOM_RENDER_SCALE;
-	vec2 stepDir = vec2(invAspectRatio, 0.0) * BLOOM_SIZE * 0.1 / SAMPLE_COUNT;
+	vec2 stepAmount = vec2(invAspectRatio, 1.0) * BLOOM_SIZE * 0.03;
+	vec3 bloomColor = vec3(0.0);
 	
-	float depth = texelFetch(DEPTH_BUFFER_ALL, texelcoord * bloomIntScale + bloomIntScale / 2, 0).r;
-	float blockDepth = toBlockDepth(depth);
-	stepDir /= blockDepth * 0.125 + 1.0;
+	doBloomTile(bloomColor, stepAmount);
+	stepAmount *= 2.0;
+	doBloomTile(bloomColor, stepAmount);
+	stepAmount *= 2.0;
+	doBloomTile(bloomColor, stepAmount);
 	
-	vec3 bloomColor = texture2DLod(BLOOM_TEXTURE, texcoord, 2).rgb;
-	for (int i = 0; i < SAMPLE_COUNT; i++) {
-		vec2 offset = stepDir * (i + 1);
-		bloomColor += texture2DLod(BLOOM_TEXTURE, texcoord + offset, 2).rgb * sampleWeights[i];
-		bloomColor += texture2DLod(BLOOM_TEXTURE, texcoord - offset, 2).rgb * sampleWeights[i];
-	}
-	bloomColor /= weightsTotal;
+	bloomColor /= weightsTotal * 3.0;
 	
 	/* DRAWBUFFERS:4 */
 	gl_FragData[0] = vec4(bloomColor, 1.0);