+++ shaders/style_vanilla.glsl
@@ -10,9 +10,17 @@
 #undef LAVA_NOISE_ENABLED
 										#define LAVA_NOISE_ENABLED 1
 #endif
-#if LAVA_NOISE_AMOUNT == -1
-#undef LAVA_NOISE_AMOUNT
-										#define LAVA_NOISE_AMOUNT 0.85
+#if LAVA_NOISE_AMOUNT_OVERWORLD == -1
+#undef LAVA_NOISE_AMOUNT_OVERWORLD
+										#define LAVA_NOISE_AMOUNT_OVERWORLD 0.3
+#endif
+#if LAVA_NOISE_AMOUNT_NETHER == -1
+#undef LAVA_NOISE_AMOUNT_NETHER
+										#define LAVA_NOISE_AMOUNT_NETHER 0.8
+#endif
+#if LAVA_NOISE_AMOUNT_END == -1
+#undef LAVA_NOISE_AMOUNT_END
+										#define LAVA_NOISE_AMOUNT_END 0.2
 #endif
 #if FANCY_NETHER_PORTAL_ENABLED == -1
 #undef FANCY_NETHER_PORTAL_ENABLED
@@ -548,15 +556,15 @@
 #endif
 #if BLOOM_NETHER_AMOUNT == -1
 #undef BLOOM_NETHER_AMOUNT
-										#define BLOOM_NETHER_AMOUNT 0.85
+										#define BLOOM_NETHER_AMOUNT 0.6
 #endif
 #if BLOOM_END_AMOUNT == -1
 #undef BLOOM_END_AMOUNT
-										#define BLOOM_END_AMOUNT 0.65
+										#define BLOOM_END_AMOUNT 0.45
 #endif
 #if BLOOM_SIZE == -1
 #undef BLOOM_SIZE
-										#define BLOOM_SIZE 1.3
+										#define BLOOM_SIZE 1.2
 #endif
 #if BLOOM_LOW_CUTOFF == -1
 #undef BLOOM_LOW_CUTOFF